% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{add_basal_disease}
\alias{add_basal_disease}
\title{Add Basal Disease}
\usage{
add_basal_disease(data, lookup)
}
\arguments{
\item{data}{A tidy data set with a `disease` column.}

\item{lookup}{A lookup table with `disease` and `nesting_disease`
columns that describe a global disease hierarchy that will be applied
to find the basal disease of each `disease` in data.}
}
\value{
tidy dataset with basal disease
}
\description{
Add column `basal_disease` to tidy dataset
}
