% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{add_column_summaries}
\alias{add_column_summaries}
\title{Add Column Summaries}
\usage{
add_column_summaries(tidy_data, dataset_name, metadata)
}
\arguments{
\item{tidy_data}{Data frame of prepared data that
are ready to be packaged as an IIDDA tidy data set.}

\item{dataset_name}{Character string giving IIDDA identifier
of the dataset.}

\item{metadata}{Output of \code{\link{get_tracking_metadata}}.}
}
\description{
Add lists of unique values and ranges of values to a
the metadata of an IIDDA data set.
}
