% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{add_filter_group_values}
\alias{add_filter_group_values}
\title{Add Filter Group Values}
\usage{
add_filter_group_values(tidy_data, dataset_name, metadata)
}
\arguments{
\item{tidy_data}{Data frame of prepared data that
are ready to be packaged as an IIDDA tidy data set.}

\item{dataset_name}{Character string giving IIDDA identifier
of the dataset.}

\item{metadata}{Output of \code{\link{get_tracking_metadata}}.}
}
\description{
Add lists of unique sets of values for a given filter group
}
