% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{add_metadata}
\alias{add_metadata}
\title{Add Metadata}
\usage{
add_metadata(table, table_metadata, column_metadata, product)
}
\arguments{
\item{table}{dataframe (or dataframe-like object)}

\item{table_metadata}{named list (or list-like object) such that
\code{table_metadata$Title} and \code{table_metadata$Description}
are strings containing the title and description of the table}

\item{column_metadata}{dataframe with rownames equal to the columns
in \code{table}, and \code{Title} and \code{Description} columns
giving the title and description of each column in \code{table}}
}
\value{
version of \code{table} with added metadata \code{attributes}
}
\description{
Add title and description metadata to a table and its columns.
}
