% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{add_user_entries}
\alias{add_user_entries}
\title{Add entries to user table}
\usage{
add_user_entries(entries, user_table_path)
}
\arguments{
\item{entries}{dataframe or named list of entries to add}

\item{user_table_path}{string indicating path to user lookup table}
}
\value{
new user lookup table with added entries at the path
}
\description{
Adds entries to user-defined lookup table
Entries should have list names or columns in the user lookup table
\code{standards} can be used within entries
}
