% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{basal_disease}
\alias{basal_disease}
\title{Basal Disease}
\usage{
basal_disease(disease, disease_lookup, encountered_diseases = character())
}
\arguments{
\item{disease}{Disease for which to determine basal disease}

\item{disease_lookup}{Table with two columns -- disease and nesting_disease}

\item{encountered_diseases}{Character vector of diseases already found.
Typically this left at the default value of an empty character vector.}
}
\value{
The root disease that input disease maps to in disease_lookup.
}
\description{
Basal Disease
}
