% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_parsing.R
\name{blob_to_raw}
\alias{blob_to_raw}
\title{Blob to Raw}
\usage{
blob_to_raw(urls)
}
\arguments{
\item{urls}{Character vector of GitHub URLs in blob storage}
}
\description{
Convert URL in GitHub blob storage format to GitHub raw data format.
}
\examples{
blob_to_raw("https://github.com/canmod/iidda-tools/blob/main/R/iidda/R/github_parsing.R")
}
