% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{cell_block}
\alias{cell_block}
\title{Cell Block}
\usage{
cell_block(cells_data)
}
\arguments{
\item{cells_data}{Data read in using `xlsx_cells`, or just any data frame
with integer columns `row` and `col`.}
}
\description{
Create a data frame for representing a rectangular range of cells
in an Excel file. This is useful for adding blank cells that do not
get read in by `xlsx_cells`.
}
