% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{check_tidy_data_cols}
\alias{check_tidy_data_cols}
\title{Error if columns in the metadata Schema are not in tidy data}
\usage{
check_tidy_data_cols(table, column_metadata)
}
\arguments{
\item{table}{dataframe (or dataframe-like object)}

\item{column_metadata}{dataframe with rownames equal to the columns
in \code{table}, and \code{Title} and \code{Description} columns
giving the title and description of each column in \code{table}}
}
\description{
Error if columns in the metadata Schema are not in tidy data
}
