% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{check_tracking_tables}
\alias{check_tracking_tables}
\title{Check Tracking Table Consistency}
\usage{
check_tracking_tables(path)
}
\arguments{
\item{path}{path to tracking tables}
}
\description{
Check Tracking Table Consistency
}
