% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic_to_new_iidda.R
\name{classic_to_new_iidda}
\alias{classic_to_new_iidda}
\title{Create New IIDDA Dataset from Classic IIDDA}
\usage{
classic_to_new_iidda(classic_file, new_repo)
}
\arguments{
\item{classic_file}{path to classic IIDDA file}

\item{new_repo}{path to new IIDDA repository}
}
\value{
No return value. Call to produce a new directory structure in a new
IIDDA git repository containing a classic IIDDA dataset as the source.
}
\description{
Create New IIDDA Dataset from Classic IIDDA
}
