% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{collapse_xlsx_value_columns}
\alias{collapse_xlsx_value_columns}
\title{Collapse xlsx Value Columns}
\usage{
collapse_xlsx_value_columns(data)
}
\arguments{
\item{data}{Data frame representing an xlsx file.}
}
\description{
Collapse all value columns into a single \code{\link{character}} column
for data frames that have one row per cell in an xlsx file.
}
