% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{combine_weeks}
\alias{combine_weeks}
\title{Combine Weeks}
\usage{
combine_weeks(cleaned_sheets, sheet_dates, metadata)
}
\arguments{
\item{cleaned_sheets}{List of data frames -- one for each sheet}

\item{sheet_dates}{Data frame describing sheet dates (TODO: more info needed)}

\item{metadata}{Output of \code{\link{get_tracking_metadata}}.}
}
\description{
Combine data from different Excel sheets associated with
specific weeks in 1956-2000 Canadian communicable disease
incidence data prep pipelines.
}
