% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{convert_metadata_path}
\alias{convert_metadata_path}
\title{Convert Metadata Path}
\usage{
convert_metadata_path(metadata_path, harmonized_source, tidy_source)
}
\arguments{
\item{metadata_path}{Path to a collection of tracking tables.}

\item{harmonized_source}{IIDDA data source ID for a harmonized source.}

\item{tidy_source}{IIDDA data source ID for a data source that is being
harmonized.}
}
\description{
Convert a metadata path to one corresponding to tidy data being harmonized.
}
