% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_diff.R
\name{cp_git_version}
\alias{cp_git_version}
\title{Copy old git File Version}
\usage{
cp_git_version(file, version_hash)
}
\arguments{
\item{file}{Path to file.}

\item{version_hash}{Git version hash.}
}
\value{
Temporary file path containing the copy.
}
\description{
Create a temporary file containing a copy of a file under git version
control for a particular revision of that file.
}
