% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{data_to_json_files}
\alias{data_to_json_files}
\title{Data to JSON Files}
\usage{
data_to_json_files(data, json_dir, name_field, use_extension = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{json_dir}{Path to the directory for saving the JSON files.}

\item{name_field}{Name of the field in the CSV file that contains
the names for each JSON file. All values in this field must be unique.}

\item{use_extension}{If there is a column in the CSV file called
`extension`, should it be used to produced json filenames of the form
`value-in-name-field.value-in-extension-field.json`?}
}
\description{
Create a directory of JSON files from a data frame.
}
