% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{disease_coverage_heatmap}
\alias{disease_coverage_heatmap}
\title{Creates a heatmap that shows disease coverage over the years}
\usage{
disease_coverage_heatmap(table, disease_col = "disease")
}
\arguments{
\item{table}{dataframe (or dataframe-like object). Tidy dataset of all compiled datasets}

\item{disease_col}{specifies level of disease (i.e. disease_family, disease, disease_subclass)}
}
\description{
Values are TRUE if that particular disease occurred at least once in a period that ended in that
particular year, and FALSE otherwise.
}
