% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{do_time_scale_cross_check}
\alias{do_time_scale_cross_check}
\title{Time Scale Cross Check}
\usage{
do_time_scale_cross_check(sum_of_timescales)
}
\arguments{
\item{sum_of_timescales}{Dataframe with aggregations to different time
scales (TODO: describe)}
}
\description{
Function used to produce the time-scale cross-check for the CANMOD
digitization project.
}
