% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tools.R
\name{drop_empty_rows}
\alias{drop_empty_rows}
\title{Drop Empty Rows}
\usage{
drop_empty_rows(table)
}
\arguments{
\item{table}{data frame}
}
\description{
Drop empty rows in a table using \code{\link{is_empty}}.
}
