% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{empty_column_report}
\alias{empty_column_report}
\title{Empty Column Report}
\usage{
empty_column_report(data, columns, dataset_id)
}
\arguments{
\item{data}{Data frame.}

\item{columns}{Character vector of columns giving the columns to
check for emptiness.}

\item{dataset_id}{ID for the dataset that \code{data} will become, likely
after further processing.}
}
\description{
Save the records of a dataset that contain empty values in
`columns`. This report will be saved in the
`supporting-output/dataset_id` directory.
}
