% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{empty_is_blank}
\alias{empty_is_blank}
\title{Empty is Blank}
\usage{
empty_is_blank(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Force empty strings to be blank. See \code{\link{is_empty}}.
}
