% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{empty_to_na}
\alias{empty_to_na}
\title{Convert all missing values to NA}
\usage{
empty_to_na(data)
}
\arguments{
\item{data}{data frame resulting from data prep scripts}
}
\description{
Convert all missing values to NA
}
