% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiweek_end_date}
\alias{epiweek_end_date}
\title{End-Dates of Epiweeks}
\usage{
epiweek_end_date(year, week)
}
\arguments{
\item{year}{Integer vector of years.}

\item{week}{Integer vector of weeks.}
}
\value{
Date vector of the end-dates of each specified epiweek.
}
\description{
End-Dates of Epiweeks
}
\examples{
epiweek_end_date(2010, 1:11)
}
