% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{extract_char_or_blank}
\alias{extract_char_or_blank}
\title{Extract Character or Blank}
\usage{
extract_char_or_blank(l, e)
}
\arguments{
\item{l}{List}

\item{e}{Name of the focal element}
}
\description{
Extract a character vector from a list or return
a blank string if it doesn't exist or if a proper
list isn't passed.
}
