% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{extract_or_blank}
\alias{extract_or_blank}
\title{Extract or Blank}
\usage{
extract_or_blank(l, e)
}
\arguments{
\item{l}{List}

\item{e}{Name of the focal element}
}
\description{
Try to extract a list element, and return a blank
list if it doesn't exist or if a proper list is not passed.
}
