% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_age_groups.R
\name{fill_re_template}
\alias{fill_re_template}
\title{Fill Regex Template}
\usage{
fill_re_template(re_template, which_bound = "neither")
}
\arguments{
\item{re_template}{template that resolve to regular expressions
for matching age information contained in category names}

\item{which_bound}{resolve the template to match lower or upper bounds,
neither (the default), or single}
}
\description{
Resolve a length-1 character vector containing a regex template into a
regular expression for matching age bound information in disease category
names
}
