% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tools.R
\name{fix_csv}
\alias{fix_csv}
\title{Fix CSV}
\usage{
fix_csv(filename)
}
\arguments{
\item{filename}{Path to the CSV file}
}
\value{
Logical value that is `TRUE` if the CSV needed fixing
and `FALSE` otherwise.
}
\description{
Fix the format of a CSV file that is not in IIDDA format.
}
