% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{freq_to_by}
\alias{freq_to_by}
\title{Frequency to By}
\usage{
freq_to_by(freq)
}
\arguments{
\item{freq}{one of `"weekly"` (becomes `"7 days"`),
`"4-weekly"` (becomes `"28 days"`),
`"monthly"` (becomes `"1 month"`)}
}
\description{
Convert words describing frequencies to phrases.
}
