% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{freq_to_days}
\alias{freq_to_days}
\title{Frequency to Days}
\usage{
freq_to_days(freq)
}
\arguments{
\item{freq}{one of `"weekly"` (becomes `7`),
`"4-weekly"` (becomes `28`),
`"monthly"` (returns an error)}
}
\description{
Convert words describing frequencies to corresponding numbers of days
}
