% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{generate_user_table}
\alias{generate_user_table}
\title{Create user-defined lookup table}
\usage{
generate_user_table(path, lookup_table_type)
}
\arguments{
\item{path}{string indicating path to directory}
}
\value{
.csv file of empty lookup table with columns from \code{lookup_table_type} in the directory if successful
}
\description{
Creates an empty user-defined lookup table in a specified directory
}
