% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{get_canmod_digitization_metadata}
\alias{get_canmod_digitization_metadata}
\title{Get CANMOD Digitization Metadata}
\usage{
get_canmod_digitization_metadata(tracking_list)
}
\arguments{
\item{tracking_list}{output of \code{read_tracking_tables}}
}
\description{
Superseded by functionality in `iidda.api`.
}
