% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{get_dataset_metadata}
\alias{get_dataset_metadata}
\title{Get Dataset Metadata}
\usage{
get_dataset_metadata(dataset)
}
\arguments{
\item{dataset}{Dataset identifier.}
}
\description{
Get an object with metadata information about a particular dataset from
tracking tables.
}
