% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{get_dataset_path}
\alias{get_dataset_path}
\title{Get Dataset path}
\usage{
get_dataset_path(source, dataset, ext = "csv")
}
\arguments{
\item{source}{Source ID.}

\item{dataset}{dataset ID.}

\item{ext}{Dataset file extension.}
}
\description{
Get Dataset path
}
