% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{get_implied_zeros}
\alias{get_implied_zeros}
\title{Get Implied Zeros}
\usage{
get_implied_zeros(data)
}
\arguments{
\item{data}{A tidy data set}
}
\value{
A tidy data set with inferred 0s
}
\description{
Add zeros to data set that are implied by a '0' reported at a coarser timescale.
}
