% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{get_items}
\alias{get_items}
\title{Get Items}
\usage{
get_items(l, keys)
}
\arguments{
\item{l}{A list of lists.}

\item{keys}{Name of the items in the inner lists.}
}
\description{
Get list of items within each inner list of a list of lists
}
