% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_lookup_table}
\alias{get_lookup_table}
\title{Get Lookup Table}
\usage{
get_lookup_table(table_name = c("location_iso"))
}
\arguments{
\item{table_name}{Name of a lookup table}
}
\description{
Get Lookup Table
}
