% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{get_main_script}
\alias{get_main_script}
\title{Get Main Script}
\usage{
get_main_script(source, dataset, root)
}
\arguments{
\item{source}{Source ID.}

\item{dataset}{dataset ID.}

\item{root}{Path to the root of the repository.}
}
\description{
Get Main Script
}
