% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{get_source_path}
\alias{get_source_path}
\title{Get Source Path}
\usage{
get_source_path(source, root)
}
\arguments{
\item{source}{Source ID.}

\item{root}{Path to the root of the repository.}
}
\description{
Get Source Path
}
