% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_unique_col_values}
\alias{get_unique_col_values}
\title{Unique Column Values}
\usage{
get_unique_col_values(l)
}
\arguments{
\item{l}{list of data frames with the same column names}
}
\value{
list of unique values in each column
}
\description{
Unique Column Values
}
