% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{get_with_key}
\alias{get_with_key}
\title{Get with Key by Regex}
\usage{
get_with_key(l, key, pattern, ...)
}
\arguments{
\item{l}{list of lists}

\item{key}{name of item in inner list}

\item{pattern}{regex pattern with which to match values of the key}

\item{...}{additional arguments to pass to \code{\link{grepl}}}
}
\value{
subset of elements of \code{l} that match the pattern
}
\description{
Get with Key by Regex
}
