% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_parsing.R
\name{git_path_to_raw_github}
\alias{git_path_to_raw_github}
\title{Convert GitHub URLs into Raw Format (not working)}
\usage{
git_path_to_raw_github(urls, branch = "master")
}
\arguments{
\item{urls}{TODO}

\item{branch}{TODO}
}
\description{
Convert GitHub URLs into Raw Format (not working)
}
