% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{group_with_dash}
\alias{group_with_dash}
\title{Simplify String with List of Numbers Grouped by Dashes}
\usage{
group_with_dash(x)
}
\arguments{
\item{x}{atomic vector}
}
\value{
length-1 character string giving a sorted list of numbers
with contiguous numbers grouped by dashes.
}
\description{
Simplify String with List of Numbers Grouped by Dashes
}
\examples{
group_with_dash(c("3840", "34", "2", "3", "1", "33", '5-50'))
group_with_dash(group_with_dash(c("3840", "34", "2", "3", "1", "33", '5-50')))
}
