% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_lookup.R
\docType{data}
\name{harmonization_lookup_tables}
\alias{harmonization_lookup_tables}
\title{Harmonization Lookup Tables}
\format{
A list of data frames, one for each column with historical
naming inconsistencies:
\describe{
  \item{location}{
    \describe{
      \item{location}{Unique names of locations found in IIDDA}
      \item{iso_3166}{National jurisdiction codes}
      \item{iso_3166_2}{Sub-national jurisdiction codes}
    }
  }
  \item{sex}{
    \describe{
      \item{sex}{Unique names of sexes found in IIDDA}
      \item{iso_5218}{Numeric sex codes}
    }
  }
}
}
\usage{
harmonization_lookup_tables
}
\description{
List of lookup tables for harmonizing historical inconsistencies
in naming.
}
\details{
For example, NFLD and Newfoundland can both be represented using
the iso-3166-2 standard as CA-NL. These tables can be joined to
data in IIDDA to produce standardized variables that harmonize
historical inconsistencies.
}
\keyword{datasets}
