% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_convert.R
\name{icd_finder}
\alias{icd_finder}
\title{ICD Finder}
\usage{
icd_finder(disease_pattern, maximum_number_results = 10L, ...)
}
\arguments{
\item{disease_pattern}{Regex pattern describing a disease.}

\item{maximum_number_results}{Integer giving the maximum number
of ICD codes to return, with preference given to shorter codes.}

\item{...}{Arguments to pass on to \code{\link{grepl}}.
It is recommended to set \code{ignore.case = TRUE} and often
\code{perl = TRUE}.}
}
\description{
Return the Shortest ICD-10 Codes that Match a Regex Pattern.
Requires an internet connection.
}
\examples{
icd_finder("chick")  ## Struc by chicken!!
}
