% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{identify_scales}
\alias{identify_scales}
\title{Identify Scales}
\usage{
identify_scales(data, location_type_fixer = canada_province_scale_finder)
}
\arguments{
\item{data}{Data frame in IIDDA tidy format to add time scale
and location scale information.}

\item{location_type_fixer}{Function that takes a data frame in IIDDA
tidy format and adds or fixes the `location_type` field.}
}
\description{
Identifies time scales (wk, mt, qrtr, yr) and location types (province or country) within a tidy dataset.
}
