% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{iidda_data_dictionary}
\alias{iidda_data_dictionary}
\title{IIDDA Data Dictionary}
\usage{
iidda_data_dictionary()
}
\description{
Get the global data dictionary for IIDDA
}
\details{
This function requires an internet connection.
}
