% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_iidda_dataset.R
\name{initialize_iidda_readme}
\alias{initialize_iidda_readme}
\title{Make IIDDA Dataset README File}
\usage{
initialize_iidda_readme(
  new_data_path,
  new_data_name,
  base_source_file,
  lifecycle
)
}
\description{
Make IIDDA Dataset README File
}
