% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{interp_pop}
\alias{interp_pop}
\title{Interpolate Population}
\usage{
interp_pop(tidy_pop)
}
\arguments{
\item{tidy_pop}{One or more tidy population data frames}
}
\description{
Interpolate Population
}
