% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_empty}
\alias{is_empty}
\title{Is Empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Return \code{\link{TRUE}} if a string is empty. Emptiness means that
any one of the following is true: \code{NA}, \code{NaN},
\code{nchar(as.character(x)) == 0L},
\code{tolower(as.character(x)) == "na"}
}
