% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{iso_3166_codes}
\alias{iso_3166_codes}
\title{ISO-3166 and ISO-3166-2 Codes}
\usage{
iso_3166_codes(tidy_data, locations_iso)
}
\arguments{
\item{tidy_data}{data frame containing a field called \code{location}
containing geographical location information extracted from a source
document}

\item{locations_iso}{table containing three columns: \code{location}
with all unique location identifiers in the \code{tidy_data},
\code{iso_3166} containing equivalent ISO-3166 codes (if applicable), and
\code{iso_3166_2} containing equivalent ISO-3166-2 codes (if applicable)}
}
\description{
Converts geographical location information, as it was described in a
source document, to equivalent ISO-3166 and ISO-3166-2 codes.
}
