% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{iso_8601_dateranges}
\alias{iso_8601_dateranges}
\title{ISO-8601 Date Ranges}
\usage{
iso_8601_dateranges(start_date, end_date)
}
\arguments{
\item{start_date}{date vector}

\item{end_date}{date vector}
}
\description{
Converts start and end dates
into ISO-8601-compliant date ranges.
}
