% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{iso_8601_dates}
\alias{iso_8601_dates}
\title{ISO-8601 Dates}
\usage{
iso_8601_dates(dates)
}
\arguments{
\item{dates}{date vector}
}
\description{
Convert date vectors into string vectors with ISO-8601
compliant format.
}
