% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{json_files_to_csv}
\alias{json_files_to_csv}
\title{JSON Files to CSV}
\usage{
json_files_to_csv(json_paths, csv_path)
}
\arguments{
\item{json_paths}{Vector of paths to JSON files.}

\item{csv_path}{Path for saving the resulting CSV file.}
}
\description{
Create a CSV file from a set of JSON files.
}
