% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{json_files_to_data}
\alias{json_files_to_data}
\title{JSON Files to Data}
\usage{
json_files_to_data(json_paths)
}
\arguments{
\item{json_paths}{Vector of paths to JSON files.}
}
\description{
Create a data frame from a set of JSON files.
}
