% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{key_val}
\alias{key_val}
\title{Key-Value}
\usage{
key_val(l, key, value)
}
\arguments{
\item{l}{A list of named lists}

\item{key}{A name of an element in each list in \code{l}}

\item{value}{A name of an element in each list in \code{l}}
}
\description{
Create a set of key-value pairs by extracting elements from
within a list of named-lists.
}
\examples{
f = system.file("example_data_dictionary.json", package = "iidda")
d = jsonlite::read_json(f)
key_val(d, "name", "type")
}
