% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_dataset_ids}
\alias{list_dataset_ids}
\title{List Dataset IDs}
\usage{
list_dataset_ids(source, root)
}
\arguments{
\item{source}{Source ID.}

\item{root}{Path to the root of the repository.}
}
\description{
List Dataset IDs
}
