% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_dataset_ids_by_source}
\alias{list_dataset_ids_by_source}
\title{List Dataset IDs by Source}
\usage{
list_dataset_ids_by_source(root)
}
\arguments{
\item{root}{Path to the root of the repository.}
}
\description{
List Dataset IDs by Source
}
