% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_dependency_ids}
\alias{list_dependency_ids}
\title{List Dependency IDs}
\usage{
list_dependency_ids(
  source,
  dataset,
  type = c("PrepScripts", "Scans", "Digitizations", "AccessScripts"),
  root
)
}
\arguments{
\item{source}{Source ID.}

\item{dataset}{Dataset ID.}

\item{type}{Type of resource.}

\item{root}{Path to the root of the repository.}
}
\description{
List Dependency IDs
}
