% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{list_extract}
\alias{list_extract}
\title{List Extract}
\usage{
list_extract(x, pattern, ...)
}
\arguments{
\item{x}{A list.}

\item{pattern}{A regular expression}

\item{...}{Arguments to pass to \code{\link{grepl}}}
}
\description{
Extract list items by regular expression matching
on their names.
}
