% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_file_id}
\alias{list_file_id}
\title{List File ID}
\usage{
list_file_id(..., ext, root)
}
\arguments{
\item{...}{Path components to directory containing the resources.}

\item{ext}{Optional string giving the file extension of the resources. If
missing then all resources are given.}

\item{root}{Optional project root.}
}
\value{
List of matching files without their extensions.
}
\description{
List File ID
}
