% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_prep_script_ids}
\alias{list_prep_script_ids}
\title{List Prep Script IDs}
\usage{
list_prep_script_ids(source, root)
}
\arguments{
\item{source}{Source ID.}

\item{root}{Path to the root of the repository.}
}
\description{
List Prep Script IDs
}
