% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_resource_ids}
\alias{list_resource_ids}
\title{List Resources IDs}
\usage{
list_resource_ids(
  source,
  type = c("TidyDatasets", "PrepScripts", "Scans", "Digitizations", "AccessScripts")
)
}
\arguments{
\item{source}{Source ID.}

\item{type}{Type of resource.}
}
\description{
List Resources IDs
}
