% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{list_xpath}
\alias{list_xpath}
\title{List XPath}
\usage{
list_xpath(l, ...)
}
\arguments{
\item{l}{A hierarchical list.}

\item{...}{Character strings describing the path down the hierarchy.}
}
\description{
Extract elements of lists using x-path-like syntax.
}
\examples{
l = list(
  a = list(
    A = list(
      i = 1,
      ii = 2
    ),
    B = list(
      i = 3,
      ii = 4
    )
 ),
 b = list(
    A = list(
      i = 5,
      ii = 6
    ),
    B = list(
      i = 7,
      ii = 8
    )
  )
)
list_xpath(l, "A", "i")
list_xpath(l, "B", "ii")
}
