% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{lookup}
\alias{lookup}
\title{Lookup Value}
\usage{
lookup(named_keys, l)
}
\arguments{
\item{named_keys}{named character vector with values giving keys
to lookup in \code{l}}

\item{l}{list with names to match against the
values of \code{keys}}
}
\description{
Lookup Value
}
