% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_age_groups.R
\name{make_age_hash_table}
\alias{make_age_hash_table}
\title{Make Age Hash Table}
\usage{
make_age_hash_table(
  categories,
  re_templates,
  which_bound = c("lower", "upper", "neither", "single"),
  prefix = ""
)
}
\arguments{
\item{categories}{character vector of disease category names}

\item{re_templates}{list of templates that resolve to regular expressions
for matching age information contained in category names}

\item{which_bound}{should the lookup function return lower or upper age
bounds}
}
\value{
vector containing either the lower or upper age bounds contained
in the categories
}
\description{
Create a lookup function that takes a character vector of disease category
names and returns a vector of equal length containing either the lower
or upper age bounds contained in the categories. If no bound is present
then \code{NA} is returned.
}
