% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_helpers.R
\name{make_compilation_dependencies}
\alias{make_compilation_dependencies}
\title{Make Compilation Dependencies}
\usage{
make_compilation_dependencies(compilation_dataset, dataset_paths)
}
\arguments{
\item{compilation_dataset}{Dataset ID for which dependencies are being
declared.}

\item{dataset_paths}{Relative paths to dependencies.}
}
\description{
Create a dependency file and prep script for a dataset that is a compilation
of other datasets. These files are created once and any edits should be
made manually to the created files.
}
