% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_config.R
\name{make_config}
\alias{make_config}
\title{Create IIDDA Config File}
\usage{
make_config(
  path = file.path(getwd(), "config.json"),
  iidda_owner = "",
  iidda_repo = "",
  github_token = "",
  .overwrite = FALSE
)
}
\arguments{
\item{path}{path for storing config file}

\item{iidda_owner}{TODO}

\item{iidda_repo}{TODO}

\item{github_token}{TODO}

\item{.overwrite}{should existing config.json files be overwritten}
}
\description{
Create IIDDA Config File
}
