% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{make_data_cite}
\alias{make_data_cite}
\title{Make DataCite JSON Metadata}
\usage{
make_data_cite(metadata, file)
}
\arguments{
\item{metadata}{Output of get_tracking_metadata}

\item{file}{Path to metadata file}
}
\description{
Make DataCite JSON Metadata
}
