% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_helpers.R
\name{make_dataset_metadata}
\alias{make_dataset_metadata}
\title{Make Dataset Metadata}
\usage{
make_dataset_metadata(tidy_dataset, type, ...)
}
\arguments{
\item{tidy_dataset}{Dataset ID for which metadata is being produced.}

\item{type}{Type of dataset (e.g., CDI, Mortality).}

\item{...}{Additional metadata fields to provide. If invalid fields are
supplied, an error message will be given.}
}
\description{
Make Dataset Metadata
}
