% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_generation.R
\name{make_generic_makefile}
\alias{make_generic_makefile}
\title{Make a Generic Makefile}
\usage{
make_generic_makefile(makefile_dir)
}
\description{
This function is for IIDDA maintainers only. It id used to generate
a makefile that recursively looks for makefiles within a folder hierarchy.
}
