% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_generation.R
\name{make_makefile}
\alias{make_makefile}
\title{Generate Dataset Makefile}
\usage{
make_makefile(source, dataset)
}
\arguments{
\item{source}{IIDDA source ID.}

\item{dataset}{IIDDA dataset ID.}
}
\description{
Run this function from the root of an IIDDA repository.
}
