% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_helpers.R
\name{make_resource_metadata}
\alias{make_resource_metadata}
\title{Make Resource Metadata}
\usage{
make_resource_metadata(source)
}
\arguments{
\item{source}{Source ID.}
}
\description{
Make one json metadata file for each resource (i.e., prep/access script or
digitization/scan of data)) in a source pipeline associated with a
data source (i.e., a sub-directory of `pipelines`). Existing metadata
files will not be overwritten.
}
