% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_helpers.R
\name{make_source_directory}
\alias{make_source_directory}
\title{Make Source Directory}
\usage{
make_source_directory(source, files)
}
\arguments{
\item{source}{Source ID.}

\item{files}{Character vector of files that are either already in the
pipeline or that should be added.}
}
\description{
Make a sub-directory of `pipelines` containing a data and/or code source.
}
