% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_helpers.R
\name{make_source_metadata}
\alias{make_source_metadata}
\title{Make Source Metadata}
\usage{
make_source_metadata(source, organization, location, ...)
}
\arguments{
\item{source}{Source ID.}

\item{organization}{Organization from which the source was obtained.}

\item{location}{Location for which data was collected.}

\item{...}{Additional metadata fields to provide. If invalid fields are
supplied, an error message will be given.}
}
\description{
Make a json file associated with a new data source (i.e., a sub-directory
of `pipelines`).
}
