% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{melt_tracking_table_keys}
\alias{melt_tracking_table_keys}
\title{Melt Tracking Table Keys (Deprecated)}
\usage{
melt_tracking_table_keys(keys)
}
\arguments{
\item{keys}{Character vector of}
}
\description{
To be used in conjunction with \code{\link{tracking_table_keys}}.
}
