% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_config.R
\name{MissingHandlers}
\alias{MissingHandlers}
\title{Missing Handlers}
\usage{
MissingHandlers(
  unclear = c("Unclear", "unclear", "uncleaar", "uncelar", "r"),
  not_reported = c("", "Not available", "*", "Not reportable"),
  zeros = c("—", "⎻", "-", "‾", "_", "–")
)
}
\arguments{
\item{unclear}{Character vector giving values corresponding to numbers
that were unclear to data enterers.}

\item{not_reported}{Character vector giving values corresponding to
numbers that were not reported in the original source.}

\item{zeros}{Character vector giving values corresponding to
`0` but that were entered as another character to resemble the original
source.}
}
\value{
An environment with functions for handling missing values.
}
\description{
Construct an object with functions for handling missing values.
}
