% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{non_numeric_report}
\alias{non_numeric_report}
\title{Non-Numeric Report}
\usage{
non_numeric_report(data, numeric_column, dataset_id)
}
\arguments{
\item{data}{Data frame.}

\item{numeric_column}{Name of a numeric column in \code{data}.}

\item{dataset_id}{ID for the dataset that \code{data} will become, likely
after further processing.}
}
\description{
Save the records of a dataset that contain non-numeric data within
a specified numeric field. This report will be saved in the
`supporting-output/dataset_id` directory.
}
