% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_normalization.R
\name{normalize_diseases}
\alias{normalize_diseases}
\title{Normalize Diseases}
\usage{
normalize_diseases(diseases)
}
\arguments{
\item{diseases}{Character vector of disease names}
}
\description{
Normalize the names of diseases to simplify the
harmonization of disease names across historical
sources.
}
