% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{open_locally}
\alias{open_locally}
\title{Open a Path on Mac OS or Windows}
\usage{
open_locally(urls, command = "open", args = character())
}
\arguments{
\item{urls}{Character vector of GitHub URLs in blob storage}

\item{command}{Command-line function to use to open the file (not
applicable on Windows systems.}

\item{args}{Additional options to pass to \code{command} (ignored on
Windows systems).}
}
\description{
Open a Path on Mac OS or Windows
}
