% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{or_pattern}
\alias{or_pattern}
\title{Or Pattern}
\usage{
or_pattern(x, at_start = TRUE, at_end = TRUE)
}
\arguments{
\item{x}{Character vector of alternative patterns.}

\item{at_start}{Match only at the start of strings.}

\item{at_end}{Match only at the end of strings.}
}
\description{
Construct regex for Boolean-or.
}
