% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paging.R
\name{pager}
\alias{pager}
\title{Pager}
\usage{
pager(page, n_per_page, rev = TRUE)
}
\arguments{
\item{page}{What page should be returned?}

\item{n_per_page}{How many entries on each page?}

\item{rev}{Should page one be at the end?}
}
\value{
Function of `x` to return the `page`th `page` of size `n_per_page`
of `x`.
}
\description{
Pager
}
