% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_api_hook.R
\name{parse_columns}
\alias{parse_columns}
\title{Parse Columns}
\usage{
parse_columns(data, data_dictionary)
}
\arguments{
\item{data}{Data frame with all character-valued columns.}

\item{data_dictionary}{List of lists giving a data dictionary.}
}
\description{
Set the types of a dataset with all character-valued columns using a
data dictionary that defines the types.
}
