% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{paste_operators}
\alias{paste_operators}
\alias{\%_\%}
\alias{\%+\%}
\alias{\%.\%}
\title{Paste Operators}
\usage{
x \%_\% y

x \%+\% y

x \%.\% y
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{
\code{x} concatenated with \code{y}
}
\description{
Syntactic sugar for common string pasting operations.
}
\details{
\describe{
  \item{\code{\%+\%}}{Paste with a blank separator, like python string concatenation}
  \item{\code{\%_\%}}{Paste with underscore separator}
  \item{\code{\%.\%}}{Paste with dot separator -- useful for adding file extensions}
}
}
\examples{
'google' \%.\% 'com'
'snake' \%_\% 'case'
}
