% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_pkg_dependencies.R
\name{pkg_versions}
\alias{pkg_versions}
\title{Package Versions}
\usage{
pkg_versions(dir, file)
}
\arguments{
\item{dir}{Directory containing the R scripts}

\item{file}{Name of R file containing the}
}
\value{
Character vector containing the version numbers of packages
loaded by `library` commands without any surrounding whitespace that
appear in `R` files (recursively) in `dir`
}
\description{
Package Versions
}
