% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{proj_path}
\alias{proj_path}
\title{Project Path}
\usage{
proj_path(...)
}
\arguments{
\item{...}{Path components for \code{\link{file.path}}.}
}
\description{
Return a path in absolute form (if that is how it is specified) or
relative to the IIDDA project root found using \code{\link{proj_root}}.
}
