% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_query_tools.R
\name{query_options}
\alias{query_options}
\title{Get field names for iidda queries}
\usage{
query_options(
  response_type = c("metadata", "columns", "data_dictionary"),
  data_type = c("Communicable Disease Incidence", "Population")
)
}
\arguments{
\item{response_type}{Character, one of `metadata`, `columns`, or `data_dictionary`}

\item{data_type}{Character, one of `cdi`, `CDI`, or `Communicable Disease
Incidence` for disease incidence time series datasets, or `pop`, `Pop`,  or
`Population` for historical demographic census datasets"}
}
\value{
A Vector of field names, or if `response_type = "data_dictionary"`, a named list
}
\description{
Get field names for iidda queries
}
\examples{

query_options(response_type = "metadata"
              , data_type = "Communicable Disease Incidence")
             
query_options(response_type = "columns"
              , data_type = "Population")

# error for disallowed searches
# query_options(response_type = "disease"
# , data_type = "Communicable Disease Incidence") 

# in some cases, abbreviations allowed
query_options(response_type = "columns"
              , data_type = "cdi") 
# data dictionary returns a named list
query_options(response_type = "data_dictionary"
              , data_type = "pop")
}
