% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_column_metadata}
\alias{read_column_metadata}
\title{Read Column Metadata}
\usage{
read_column_metadata(dataset, pattern)
}
\arguments{
\item{dataset}{IIDDA dataset ID.}

\item{pattern}{Regular expression pattern for filtering candidate paths
to be read from.}
}
\description{
Read Column Metadata
}
