% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tools.R
\name{read_data_columns}
\alias{read_data_columns}
\title{Read Data Columns}
\usage{
read_data_columns(filename)
}
\arguments{
\item{filename}{Path to a CSV file in IIDDA format.}
}
\description{
Read Data Columns
}
