% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tools.R
\name{read_data_frame}
\alias{read_data_frame}
\title{Read Data Frame}
\usage{
read_data_frame(filename, col_classes = "character")
}
\arguments{
\item{filename}{String giving the filename.}

\item{col_classes}{See \code{colClasses} from \code{\link{read.table}}.}
}
\description{
Read in a data frame from a CSV file using the CSV dialect
adopted by IIDDA.
}
