% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{read_digitized_data}
\alias{read_digitized_data}
\title{Read Digitized Data}
\usage{
read_digitized_data(metadata)
}
\arguments{
\item{metadata}{Output of \code{\link{get_tracking_metadata}}.}
}
\description{
Read in digitized data to be prepared within the IIDDA project.
}
