% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_global_metadata}
\alias{read_global_metadata}
\title{Read Global Metadata}
\usage{
read_global_metadata(
  id,
  type = c("columns", "organization", "sources", "tidy-datasets")
)
}
\arguments{
\item{id}{ID to the `type` of entity.}

\item{type}{Type of entity.}
}
\description{
Read Global Metadata
}
