% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_lookup}
\alias{read_lookup}
\title{Read Lookup}
\usage{
read_lookup(lookup_id)
}
\arguments{
\item{lookup_id}{IIDDA ID associated with an item in a `lookup-tables`
directory in an IIDDA repository.}
}
\description{
Read Lookup
}
