% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_prerequisite_data}
\alias{read_prerequisite_data}
\title{Read Prerequisite Data}
\usage{
read_prerequisite_data(dataset_id, numeric_column_for_report = NULL)
}
\arguments{
\item{dataset_id}{IIDDA dataset ID.}

\item{numeric_column_for_report}{Optional numeric column name to specify
for producing a report using \code{\link{non_numeric_report}}.}
}
\description{
Read Prerequisite Data
}
