% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_prerequisite_metadata}
\alias{read_prerequisite_metadata}
\title{Read Prerequisite Metadata}
\usage{
read_prerequisite_metadata(dataset, pattern)
}
\arguments{
\item{dataset}{IIDDA dataset ID.}

\item{pattern}{Regular expression pattern for filtering candidate
paths to metadata.}
}
\description{
Read Prerequisite Metadata
}
