% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_prerequisite_paths}
\alias{read_prerequisite_paths}
\title{Read Prerequisite Paths}
\usage{
read_prerequisite_paths(dataset, pattern)
}
\arguments{
\item{dataset}{IIDDA dataset ID.}

\item{pattern}{Regular expression pattern for filtering candidate paths
to be read from.}
}
\description{
Read Prerequisite Paths
}
