% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_support.R
\name{read_resource_metadata}
\alias{read_resource_metadata}
\title{Read Resource Metadata}
\usage{
read_resource_metadata(dataset, pattern)
}
\arguments{
\item{dataset}{IIDDA dataset ID.}

\item{pattern}{Regular expression pattern for filtering candidate paths
to be read from.}
}
\description{
Read Resource Metadata
}
