% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{read_tracking_tables}
\alias{read_tracking_tables}
\title{Read Tracking Tables}
\usage{
read_tracking_tables(path)
}
\arguments{
\item{path}{Path containing tracking tables.}
}
\description{
Read metadata tracking tables for an IIDDA project.
}
