% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme_templates.R
\docType{data}
\name{readme_classic_iidda}
\alias{readme_classic_iidda}
\title{README File Template}
\format{
An object of class \code{character} of length 1.
}
\usage{
readme_classic_iidda
}
\description{
(Deprecated)
}
\keyword{datasets}
