% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_helpers.R
\name{register_prep_script}
\alias{register_prep_script}
\title{Register Prep Script}
\usage{
register_prep_script(script_path, type)
}
\arguments{
\item{script_path}{Path to the prep-script being registered.}

\item{type}{Type of the dataset being produced (e.g., CDI, Mortality).
TODO: Give a list of acceptable values. Should be programmatically produced.}
}
\description{
Convenience function for a one-time setup of all metadata required for
a new prep script. The assumptions are that (1) the prep script is a `.R`
file in the `prep-scripts` directory of a directory within the `pipelines`
directory and (2) that this script produces a csv file in the
`derived-datasets` directory with the same `basename()` as this `.R` file.
Messages are printed with paths to newly created and/or existing metadata,
derived data, and dependency files that should be checked manually. Sometimes
it is helpful to delete some of these files and rerun `register_prep_script`.
However, this `register_prep_script` function should not be used in a
script that is intended to be run multiple times, as going forward the
metadata and dependency files should be edited manually.
}
