% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{relative_paths}
\alias{relative_paths}
\title{Relative Paths}
\usage{
relative_paths(paths, containing_path = proj_root())
}
\arguments{
\item{paths}{Vector of absolute paths.}

\item{containing_path}{Target working directory to be relative to.}
}
\description{
Convert a set of absolute paths to relative paths with respect to a
specified `containing_path`
}
