% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_age_groups.R
\name{remove_age}
\alias{remove_age}
\alias{memoise_remove_age}
\title{Remove Age}
\usage{
remove_age(categories, re_templates, prefix = "")

memoise_remove_age(categories, re_templates, prefix = "")
}
\arguments{
\item{categories}{vector of category names}

\item{re_templates}{list of templates that resolve to regular expressions
for matching age information contained in category names}
}
\description{
Remove age information from a vector of category names
}
