% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{remove_between_paren}
\alias{remove_between_paren}
\title{Remove Parenthesized Substring}
\usage{
remove_between_paren(
  x,
  left = "\\\\(",
  right = "\\\\)",
  contents_pattern = ".*"
)
}
\arguments{
\item{x}{Character vector}

\item{left}{Left parenthetical string}

\item{right}{Right parenthetical string}
}
\value{
Version of \code{x} with first parenthesized substrings removed
}
\description{
Remove Parenthesized Substring
}
\examples{
x = c("-", "", NA, "1", "3", "1 (Alta.)", "(Sask) 20")
remove_between_paren(x)
}
