% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_age_groups.R
\name{return_matched_age_bound}
\alias{return_matched_age_bound}
\title{Matched Age Bound}
\usage{
return_matched_age_bound(x)
}
\arguments{
\item{x}{character vector from the list output of regmatches,
containing regex matches of age bound information contained in
disease category names. each \code{x} corresponds to a single
category.}
}
\description{
Process output from regmatches to return the correct age bound.
Used in the lookup function created by \code{make_age_hash_table}
}
