% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rm_trailing_slash}
\alias{rm_trailing_slash}
\title{Remove Trailing Slash}
\usage{
rm_trailing_slash(x)
}
\arguments{
\item{x}{Character vector with paths}
}
\value{
Character vector without trailing slash
}
\description{
Remove Trailing Slash
}
