% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{save_result}
\alias{save_result}
\title{Save Results of a Data Prep Script}
\usage{
save_result(result, metadata)
}
\arguments{
\item{result}{Named list of data resulting from data prep scripts}

\item{metadata}{Nested named list describing metadata for the result.
It must have a \code{$Product[["Path to tidy data"]]} component, which is
a GitHub URL describing the ultimate location of the R data file.
The GitHub component of the URL will be removed to produce
a path that will correspond to the location within a cloned git
repository -- note that the path is relative to the top-level of
the cloned repository.}
}
\description{
Save the resulting objects of a data prep script into an R data file.
The names of the resulting objects are given by the names of the
result list.
}
