% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_term_ocr_functions.R
\name{school_term_ocr}
\alias{school_term_ocr}
\title{OCR Extraction of Canadian School Term Data from PDF Files}
\usage{
school_term_ocr(
  pdf_image,
  page,
  trim_top = 570,
  trim_bottom = 330,
  edge_geometry = "10x10+20\%+50\%"
)
}
\arguments{
\item{pdf_image}{result of image_read on character string pointing
to an EdCan PDF file containing school term data}

\item{page}{integer giving the page number of the PDF file containing
a school term data table}

\item{trim_top}{number of pixels at the top of the page to trim off}

\item{trim_bottom}{number of pixels at the bottom of the page to trim off}

\item{edge_geometry}{parameters to pass to the imagemagick edge detector}
}
\description{
OCR Extraction of Canadian School Term Data from PDF Files
}
