% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{set_ext}
\alias{set_ext}
\title{Set Extension}
\usage{
set_ext(paths, ext)
}
\arguments{
\item{paths}{Character vector giving file paths.}

\item{ext}{String giving the file extension to add to the \code{paths}.}
}
\description{
Set Extension
}
