% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tools.R
\name{set_iidda_col_types}
\alias{set_iidda_col_types}
\title{Set IIDDA Column Types}
\usage{
set_iidda_col_types(data)
}
\arguments{
\item{data}{Dataset from IIDDA Api}
}
\description{
Deprecated -- iidda.api package is not more robust.
}
