% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_types}
\alias{set_types}
\title{Set Data Frame Column Types}
\usage{
set_types(data, types)
}
\arguments{
\item{data}{data frame}

\item{types}{dict-like list with keys giving column names and
values giving types}
}
\value{
data frame with changed column types -- note that the
returned data frame is a plain base R \code{data.frame}
(i.e. not a \code{tibble} or \code{data.table}).
}
\description{
Set the types of the columns of a data frame.
}
