% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{source_from_digitization_id}
\alias{source_from_digitization_id}
\title{Source from Digitization ID}
\usage{
source_from_digitization_id(digitization_ids)
}
\arguments{
\item{digitization_ids}{Character vector of digitization IDs}
}
\value{
Character vector of source IDs associated with digitization.
}
\description{
Source from Digitization ID
}
