% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{split_tidy_data}
\alias{split_tidy_data}
\title{Split Tidy Data}
\usage{
split_tidy_data(tidy_data)
}
\description{
Creates 6 tidy data sets with no duplicate data, broken down by available time-series scales
(week, month, quarter, year) and by Province/Canada.
}
