% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sprintf_named}
\alias{sprintf_named}
\title{Lightweight Templating}
\usage{
sprintf_named(template, ...)
}
\description{
Version of the \code{sprintf} base R function that
adds basic templating -- https://stackoverflow.com/a/55423080/2047693
}
\examples{
sprintf_named("You might like to download datasets from \%{repo}s.", repo = "IIDDA")
}
