% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_lookup.R
\docType{data}
\name{standards}
\alias{standards}
\title{Standards}
\format{
List of lists of character vectors containing the original historical
names:
\describe{
  \item{location}{
    \describe{
      \item{iso_3166}{
        Historical national names associated with each iso-3166 code.
      }
      \item{iso_3166_inclusive}{
        Historical national and sub-national names associated with
        each iso-3166 code.
      }
      \item{iso_3166_2}{
        Historical sub-national names associated with each iso-3166-2 code.
      }
    }
  }
  \item{sex}{
    \describe{
      \item{iso_5218}{
        Historical names referring to sexes associated with
        each iso-5218 code.
      }
    }
  }
}
}
\usage{
standards
}
\description{
List of lists of lists that exploits tab completion to make it convenient
to get vectors of all synonyms associated with a particular standard code.
This mechanism is useful when searching for data in IIDDA.
}
\keyword{datasets}
