% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{statcan_mort_prep}
\alias{statcan_mort_prep}
\title{Prepare Mortality Data from Statistics Canada}
\usage{
statcan_mort_prep(data)
}
\arguments{
\item{data}{Output of \code{\link{read_digitized_data}} that has been
filtered to include only the cell range that contains data.}
}
\value{
Data frame complying with the IIDDA requirements for
tidy datasets.
}
\description{
Prepare Mortality Data from Statistics Canada
}
