% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_parsing.R
\name{strip_blob_github}
\alias{strip_blob_github}
\title{Strip Blob}
\usage{
strip_blob_github(urls)
}
\arguments{
\item{urls}{Character vector of GitHub URLs in blob storage}
}
\description{
Strip the 'blob part' of a GitHub URL so that it is
a path relative to a local clone of the associated repo.
}
\examples{
strip_blob_github("https://github.com/canmod/iidda-tools/blob/main/R/iidda/R/github_parsing.R")
}
