% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{sum_timescales}
\alias{sum_timescales}
\title{Sum Timescales}
\usage{
sum_timescales(data, filter_out_bad_time_scales = TRUE)
}
\arguments{
\item{data}{Data frame to aggregate to different time scales.}

\item{filter_out_bad_time_scales}{Should time scales be filtered out
if they do not have a reasonable number of periods (e.g., 600 weeks
in a year would be filtered out).}
}
\description{
Sum Timescales
}
