% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{summarise_dates}
\alias{summarise_dates}
\title{Summarise Dates}
\usage{
summarise_dates(x_start, x_end, range_operator = " to ", collapse = TRUE)
}
\arguments{
\item{x_start}{vector of starting dates.}

\item{x_end}{vector of ending dates.}

\item{range_operator}{string to go between the start and end date, defaults to " to ".}

\item{collapse}{boolean to collapse all dates into one comma separated string, defaults to TRUE.}
}
\value{
vector or single string of summarised date ranges.
}
\description{
Consecutive or overlapping date ranges are summarised into a single date range, non-consecutive date
ranges are kept as is.
}
