% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{summarise_integers}
\alias{summarise_integers}
\title{Summarise Integers}
\usage{
summarise_integers(x, range_operator = "-", collapse = TRUE)
}
\arguments{
\item{x}{vector of integers}

\item{range_operator}{string to go between the starting and ending integer in the range, defaults to "-".}

\item{collapse}{boolean to collapse all integer ranges into one comma separated string, defaults to TRUE.}
}
\value{
vector or single string of summarised integer ranges.
}
\description{
Consecutive or overlapping integers separated by commas or semi-colons are summarised into a single integer range, 
non-consecutive integer ranges are kept as is.
}
