% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{summarise_locations}
\alias{summarise_locations}
\title{Summarise Locations}
\usage{
summarise_locations(data)
}
\arguments{
\item{data}{Data frame hopefully containing at least one of  `iso_3166`,
`iso_3166_2`, or `location`. If all are missing then the output summary is
a blank string.}
}
\value{
A string summarizing the data in the columns.
}
\description{
Summarise several columns in an IIDDA dataset that specify the geographic
location of each row.
}
