% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{summarise_strings}
\alias{summarise_strings}
\title{Summarise Strings}
\usage{
summarise_strings(x, sep = ", ")
}
\arguments{
\item{x}{vector}

\item{sep}{character separator, defaults to ", "}
}
\value{
single string of summarised strings.
}
\description{
Summarise vector of strings separated by commas or semi-colons into a single
character separated string. Removes empty strings, repeated strings and trims white space.
}
