% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{test_result}
\alias{test_result}
\title{Test Results}
\usage{
test_result(result)
}
\arguments{
\item{result}{Named list of data resulting from data prep scripts}
}
\description{
Test the results of a data prep script (not finished).
}
