% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tools.R
\name{time_series_islands}
\alias{time_series_islands}
\title{Time Series Islands}
\usage{
time_series_islands(data, series_variable, time_variable = NULL)
}
\arguments{
\item{data}{A dataset (must be ordered if `time_variable` is `NULL`).}

\item{series_variable}{Name of a series variable.}

\item{time_variable}{Optional variable to use for ordering the dataset
before islands are located.}
}
\description{
Find 'island rows' in a dataset with ordered rows. Islands have a
series variable that is not `NA` surrounded by `NA` values in that
same variable.
}
