% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_query_tools.R
\name{token_matcher}
\alias{token_matcher}
\title{Obtain tokens matching regex strings}
\usage{
token_matcher(strings, tokens)
}
\arguments{
\item{strings}{Character vector of regex strings to match}

\item{tokens}{Character vector of tokens to match against}
}
\value{
Charaacter vector of unique matching tokens
}
\description{
Obtain tokens matching regex strings
}
\examples{
token_matcher(strings =c(".*sis.*", ".*itis.*")
    , tokens = unique_field_entries(response_type = "columns"
                               , metadata_search = "disease"
                               , string_comparison = "contains"))
}
