% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\docType{data}
\name{tracking_table_keys}
\alias{tracking_table_keys}
\title{Tracking Table Keys}
\format{
An object of class \code{list} of length 5.
}
\usage{
tracking_table_keys
}
\description{
Tracking Table Keys
}
\keyword{datasets}
