% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{tracking_tables_with_column}
\alias{tracking_tables_with_column}
\title{Which Tracking Tables have a Particular Column}
\usage{
tracking_tables_with_column(metadata, col_nm)
}
\arguments{
\item{metadata}{Output of \code{\link{read_tracking_tables}}.}

\item{col_nm}{Name of a column.}
}
\description{
Which Tracking Tables have a Particular Column
}
