% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_normalization.R
\name{two_field_format}
\alias{two_field_format}
\title{Two Field Format}
\usage{
two_field_format(dataset)
}
\arguments{
\item{dataset}{A tidy data set with `disease{|_subclass|_family}` columns.}
}
\description{
Attempt to automatically convert a dataset from `disease{|_subclass|_family}`
format of disease ID to the `{|nesting_}disease` format.
}
