% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_query_tools.R
\name{unique_field_entries}
\alias{unique_field_entries}
\title{Query unique tokens for a field in iidda metadata}
\usage{
unique_field_entries(
  response_type = c("metadata", "columns", "data_dictionary"),
  metadata_search = "disease",
  string_comparison = "contains"
)
}
\arguments{
\item{response_type}{Character, one of `metadata`, `columns`, or `data_dictionary`}

\item{metadata_search}{Character, name of a metadata field}

\item{string_comparison}{Character, one of `"exact"` or `"contains"`}
}
\value{
Character vector of unique tokens for a given field from all iidda datasets
}
\description{
Query unique tokens for a field in iidda metadata
}
\examples{

unique_field_entries(response_type = "columns"
                     , metadata_search = "disease"
                     , string_comparison = "contains")
                     
}
