% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{unlist_char_list}
\alias{unlist_char_list}
\title{Unlist a List of Character Vectors}
\usage{
unlist_char_list(x)
}
\arguments{
\item{x}{list of character vectors}
}
\description{
Replacing list elements with \code{list('')}
for each element that is null, not a character
vector, or length zero.
}
