% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile_generation.R
\name{update_makefiles}
\alias{update_makefiles}
\title{Update all Makefiles}
\usage{
update_makefiles(
  dataset_makefiles = TRUE,
  source_makefiles = FALSE,
  derived_data_makefile = FALSE
)
}
\arguments{
\item{dataset_makefiles}{Should makefiles for each dataset be constructed
using information in the `tracking` folder for that dataset?}

\item{source_makefiles}{Should generic makefiles for each source be
constructed?}

\item{derived_data_makefile}{Should an overall generic makefile for all
sources be constructed?}
}
\description{
Run this after updating CSV files in `pipelines/{DATA-SOURCE}/tracking`
folders and/or in `global-metadata/tracking`. To update the makefile for
a single dataset please use \code{\link{make_makefile}}.
}
