% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{vsub}
\alias{vsub}
\title{Vectorized String Substitution}
\usage{
vsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern, replacement, x}{first three arguments to \code{sub},
but the first is allowed to be a vector}

\item{...}{additional arguments to pass on to \code{sub}}
}
\description{
Vectorized String Substitution
}
