% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_age_groups.R
\name{wrap_age_patterns}
\alias{wrap_age_patterns}
\title{Wrap Age Patterns}
\usage{
wrap_age_patterns(
  patterns,
  purpose = c("extraction", "removal", "validate"),
  prefix = ""
)
}
\arguments{
\item{patterns}{vector of regular expressions for matching age bound
information in disease category names}

\item{purpose}{character string indicating the purpose of the resulting
regular expression}

\item{prefix}{pattern to match at the beginning of the string that marks
the beginning of age information}
}
\description{
Wrap list of regular expressions for matching age bounds in disease
category names, so that the resulting regular expressions can be used
for different purposes (extraction, removal, or validation)
}
