% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tools.R
\name{write_data_frame}
\alias{write_data_frame}
\title{Write Data Frame}
\usage{
write_data_frame(data, filename)
}
\arguments{
\item{data}{A data frame to write}

\item{filename}{string giving the filename}
}
\description{
Write a data frame to a CSV file using the CSV dialect
adopted by IIDDA.
}
