% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{write_local_data_dictionaries}
\alias{write_local_data_dictionaries}
\title{Write Local Data Dictionaries}
\usage{
write_local_data_dictionaries(metadata, path)
}
\arguments{
\item{metadata}{Output of \code{\link{read_tracking_tables}}.}

\item{path}{Path to a new JSON file.}
}
\description{
Write Local Data Dictionaries
}
