% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_diff.R
\name{xlsx_to_csv}
\alias{xlsx_to_csv}
\title{Excel to CSV}
\usage{
xlsx_to_csv(xlsx_path, csv_path)
}
\arguments{
\item{xlsx_path}{Path to an Excel file.}

\item{csv_path}{Path to a new CSV file.}
}
\description{
Convert an Excel file to a CSV file.
}
