% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{all_prep_script_outcomes}
\alias{all_prep_script_outcomes}
\alias{successful_prep_script_outcomes}
\alias{failed_prep_script_outcomes}
\alias{error_tar}
\title{Prep Script Outcomes}
\usage{
all_prep_script_outcomes()

successful_prep_script_outcomes()

failed_prep_script_outcomes()

error_tar(tar_name)
}
\arguments{
\item{tar_name}{Name of a tar archive to be created with log files of failed
prep script outcomes.}
}
\value{
Data frame with all prep script outcomes in the project.
}
\description{
Prep Script Outcomes
}
\section{Functions}{
\itemize{
\item \code{successful_prep_script_outcomes()}: Data frame with all successful prep
script outcomes

\item \code{failed_prep_script_outcomes()}: Data frame with all failed prep
script outcomes

\item \code{error_tar()}: Tar archive with log files of failed
prep script outcomes.

}}
