% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{basal_disease}
\alias{basal_disease}
\title{Basal Disease}
\usage{
basal_disease(disease, disease_lookup, encountered_diseases = character())
}
\arguments{
\item{disease}{disease for which to determine basal disease}

\item{disease_lookup}{table with two columns -- disease and nesting_disease}
}
\value{
The root disease that input disease maps to in disease_lookup.
}
\description{
Basal Disease
}
