% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_between_paren}
\alias{extract_between_paren}
\title{Extract Substring Between Parentheses}
\usage{
extract_between_paren(x, left = "\\\\(", right = "\\\\)")
}
\arguments{
\item{x}{Character vector}

\item{left}{Left parenthetical string}

\item{right}{Right parenthetical string}
}
\value{
Character vector with NA's for elements in \code{x} that
do not have parentheses and the substring between the first matching
parentheses.
}
\description{
Extract Substring Between Parentheses
}
