% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_between_paren}
\alias{extract_between_paren}
\alias{extract_all_between_paren}
\title{Extract Substring Between Parentheses}
\usage{
extract_between_paren(
  x,
  left = "\\\\(",
  right = "\\\\)",
  contents_pattern = ".*"
)

extract_all_between_paren(
  x,
  left = "\\\\(",
  right = "\\\\)",
  contents_pattern = ".*",
  max_iters = 100
)
}
\arguments{
\item{x}{Character vector}

\item{left}{Left parenthetical string}

\item{right}{Right parenthetical string}

\item{contents_pattern}{Regex pattern for the contents between parentheses}

\item{max_iters}{maximum number of items to return}
}
\value{
Character vector with NA's for elements in \code{x} that
do not have parentheses and the substring between the first matching
parentheses.
}
\description{
Note that unless you specify an appropriate \code{contents_pattern}
\code{extract_between_paren} will not work as you probably expect
if there are multiple sets of parentheses. You can use exclusion patterns
to make this work better (e.g. \code{content_pattern = '[^)]*'}).
}
\examples{
x = c("-", "", NA, "1", "3", "1 (Alta.)", "(Sask) 20")
extract_between_paren(x)
}
