% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_age_groups.R
\name{fill_and_wrap}
\alias{fill_and_wrap}
\title{Fill Template and Wrap the Results}
\usage{
fill_and_wrap(re_templates, which_bound, purpose, prefix = "")
}
\description{
Convenience function to do \code{fill_re_template} and
\code{wrap_age_patterns} in one step
}
