% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{get_all_dependencies}
\alias{get_all_dependencies}
\title{Get all Dependencies}
\usage{
get_all_dependencies(source, dataset, root)
}
\arguments{
\item{source}{Source ID.}

\item{dataset}{dataset ID.}

\item{root}{Path to the root of the repository.}
}
\description{
Get all Dependencies
}
