% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{get_tracking_metadata}
\alias{get_tracking_metadata}
\title{Read Tracking Metadata}
\usage{
get_tracking_metadata(
  tidy_dataset,
  digitization,
  tracking_path,
  original_format = TRUE
)
}
\arguments{
\item{tidy_dataset}{key to the tidy dataset being produced by the script}

\item{digitization}{key to the digitization being used by the script}

\item{original_format}{should the original tracking table format be used?}
}
\description{
Read in CSV files that contain the single-source-of-truth for metadata
to be used in a data prep script.
}
\details{
This function currently assumes that a single tidy dataset is being
produced from a single digitized file.
}
