% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_convert.R
\name{icd_finder}
\alias{icd_finder}
\title{Return the Shortest ICD-10 Codes that Match a Regex Pattern}
\usage{
icd_finder(disease_pattern)
}
\arguments{
\item{disease_pattern}{Regex pattern describing a disease}
}
\description{
Requires an internet connection
}
