% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{identify_scales}
\alias{identify_scales}
\title{Identify Scales}
\usage{
identify_scales(data)
}
\arguments{
\item{data}{data frame in IIDDA tidy format to add time scale
and location scale information

Identifies time scales (wk, mt, qrtr, yr) and location scales (prov or can) within a tidy dataset.}
}
\description{
Identify Scales
}
