% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{identify_scales}
\alias{identify_scales}
\title{Identify Scales}
\usage{
identify_scales(tidy_data)
}
\description{
Identifies time scales (wk, mt, qrtr, yr) and location scales (prov or can) within a tidy dataset.
}
