% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_iidda_dataset.R
\name{iidda_from_single_file}
\alias{iidda_from_single_file}
\title{Create New IIDDA Dataset from Single File}
\usage{
iidda_from_single_file(single_file, new_repo, lifecycle)
}
\arguments{
\item{single_file}{path to single data file}

\item{new_repo}{path to new IIDDA repository}

\item{lifecycle}{character vector giving the lifecycle state
(https://github.com/davidearn/iidda/blob/main/LIFECYCLE.md)
Probably 'Unreleased', but it could in
principle be 'Static', 'Dynamic', or 'Superseded'.}
}
\value{
No return value. Call to produce a new directory structure in a new
IIDDA git repository containing a single source data file.
}
\description{
Create New IIDDA Dataset from Single File
}
