% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iidda_from_single_file.R
\name{iidda_from_single_file}
\alias{iidda_from_single_file}
\title{Create New IIDDA Dataset from Single File}
\usage{
iidda_from_single_file(classic_file, new_repo)
}
\arguments{
\item{classic_file}{path to single data file}

\item{new_repo}{path to new IIDDA repository}
}
\value{
No return value. Call to produce a new directory structure in a new
IIDDA git repository containing a single source data file.
}
\description{
Create New IIDDA Dataset from Single File
}
