% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{iso_codes}
\alias{iso_codes}
\title{Adds ISO-3166 and ISO-3166-2 columns to tidydata}
\usage{
iso_codes(tidy_data, locations_iso = read.csv("tracking/locations_ISO.csv"))
}
\arguments{
\item{tidy_data}{tidy data.frame resulting from data prep scripts}

\item{locations_iso}{table containing all unique locations in all 
tidy data along with corresponding ISO-3166 and ISO-3166-2 codes}
}
\description{
Adds ISO-3166 and ISO-3166-2 columns to tidydata
}
