% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{list_dependency_paths}
\alias{list_dependency_paths}
\title{List Dependency Paths}
\usage{
list_dependency_paths(
  source,
  dataset,
  type = c("PrepScripts", "Scans", "Digitizations", "AccessScripts")
)
}
\arguments{
\item{source}{Source ID.}

\item{dataset}{dataset ID.}

\item{type}{Type of resource.}
}
\description{
List Dependency Paths
}
