% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_paths.R
\name{pipeline_exploration_starter}
\alias{pipeline_exploration_starter}
\title{Pipeline Exploration Quick-Start}
\usage{
pipeline_exploration_starter(script_filename, exploration_project_path, ...)
}
\arguments{
\item{script_filename}{Name for the generated script.}

\item{exploration_project_path}{Path to the folder for containing the script.
If this path doesn't exist, then it is created. If \code{script_filename}
exists in \code{exploration_project_path}, an error is returned.}

\item{...}{Additional arguments to pass to \code{\link{file.copy}}. A
useful argument here is `overwrite`, which indicates whether an existing
exploration script should be overwritten.}
}
\description{
Create an R script providing a place to start when exploring an IIDDA
pipeline.
}
\details{
The R script has the following:

1. Example code for printing out the data sources and datasets
in the IIDDA pipeline repository.
2. Code for finding the paths to datasets and to the scripts for generating
them.
3. Code for generating and/or reading in a user-selected IIDDA
dataset.

Once the data are read in, the user is free to do whatever they want to
with it.
}
