% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_parsing.R
\name{raw_github}
\alias{raw_github}
\title{Construct an URL to Download Single Files from GitHub}
\usage{
raw_github(owner, repo, path, user = NULL, token = NULL, branch = "master")
}
\arguments{
\item{owner}{User or Organization of the repo}

\item{repo}{Repository name}

\item{path}{Path to the file that you want to obtain}

\item{user}{Your username (only required for private repos)}

\item{token}{OAuth personal access token (only required for private repos)}

\item{branch}{Name of the branch (defaults to 'master')}
}
\description{
Uses the Raw GitHub API
}
