% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{rm_trailing_slash}
\alias{rm_trailing_slash}
\alias{rm_leading_slash}
\title{Remove Trailing / Leading Slash}
\usage{
rm_trailing_slash(x)

rm_leading_slash(x)
}
\arguments{
\item{x}{Character vector with paths.}
}
\value{
Character vector without trailing/leading slash.
}
\description{
Remove Trailing / Leading Slash
}
