% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{summarise_periods}
\alias{summarise_periods}
\title{Summarise Periods}
\usage{
summarise_periods(data, cutoff = 50)
}
\arguments{
\item{data}{Data frame hopefully containing both `period_start_date` and
`period_end_date`. If either are missing an error results.}

\item{cutoff}{Number of characters, above which the output string
takes the form `{max-date} to {min-date} (with gaps)`.}
}
\value{
A string summarizing the data in the columns
}
\description{
Summarise time periods in an IIDDA dataset.
}
