% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_tools.R
\name{write_tidy_data}
\alias{write_tidy_data}
\title{Write Tidy Digitized Data and Metadata}
\usage{
write_tidy_data(tidy_data, metadata)
}
\arguments{
\item{tidy_data}{Data frame of prepared data that
are ready to be packaged as an IIDDA tidy data set.}

\item{metadata}{Output of \code{\link{get_tracking_metadata}}.}
}
\value{
file names where data were written
}
\description{
Write Tidy Digitized Data and Metadata
}
